Add `collections` module
